refactor(assemble): extract token parsing code to `token.cpp`
